Update RSA.cs
Ligne 97, j'ai viré le cast en "byte" sur la lecture du fichier entrant, maintenant il ramasse le caractère quelle que soit sa valeur, et les accents peuvent enfin passer la manip :D
EZ

+++ RSA.cs
@@ -94,7 +94,7 @@ namespace RSA
                     while (sr.Peek() >= 0)
                     {
                         BigInteger i = 0;
-                        nombre = (byte)sr.Read();
+                        nombre = sr.Read();
 
                         nombre = modpow(nombre, E, N);
